+++ fdb-record-layer-lucene/src/main/java/com/apple/foundationdb/record/lucene/directory/FDBDirectoryManager.java
@@ -30,6 +30,7 @@ import com.apple.foundationdb.subspace.Subspace;
 import com.apple.foundationdb.tuple.Tuple;
 import com.apple.foundationdb.tuple.TupleHelpers;
 import com.google.common.annotations.VisibleForTesting;
+import org.apache.lucene.index.IndexOptions;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.IndexWriter;
 import org.apache.lucene.search.suggest.analyzing.AnalyzingInfixSuggester;
@@ -117,14 +118,14 @@ public class FDBDirectoryManager implements AutoCloseable {
     public AnalyzingInfixSuggester getAutocompleteSuggester(@Nullable Tuple groupingKey,
                                                             @Nonnull LuceneAnalyzerWrapper indexAnalyzerWrapper,
                                                             @Nonnull LuceneAnalyzerWrapper queryAnalyzerWrapper,
-                                                            boolean highlight) throws IOException {
+                                                            boolean highlight, @Nullable IndexOptions indexOptions) throws IOException {
         // The auto complete suggester reads and writes from a separate directory from the main
         // directory used for Lucene indexes, so add a suffix to the groupingKey to separate it
         // from the other Lucene index data but so that the data are still prefixed by the grouping key,
         // which is necessary for range deletes
         Tuple autoCompleteKey = groupingKey == null ? AUTO_COMPLETE_SUFFIX : groupingKey.addAll(AUTO_COMPLETE_SUFFIX);
         return getDirectoryWrapper(autoCompleteKey)
-                .getAutocompleteSuggester(indexAnalyzerWrapper, queryAnalyzerWrapper, highlight);
+                .getAutocompleteSuggester(indexAnalyzerWrapper, queryAnalyzerWrapper, highlight, indexOptions);
     }
 
     @Nonnull